refactor: add directory utils. move controller in this directory from components
