+++ jsconfig.json
@@ -1,14 +1,15 @@
 {
-    "compilerOptions": {
-      "baseUrl": "./src",
-      "paths": {
-        "@constants": ["./constants"],
-        "@actions": ["./store/actions"],
-        "@command": ["./command"],
-        "@components": ["./components"],
-        "@store": ["./store"],
-        "@styles": ["./styles"]
-      }
-    },
-    "exclude": ["node_modules"]
-  }
+  "compilerOptions": {
+    "baseUrl": "./src",
+    "paths": {
+      "@constants": ["./constants"],
+      "@actions": ["./store/actions"],
+      "@command": ["./command"],
+      "@components": ["./components"],
+      "@store": ["./store"],
+      "@styles": ["./styles"],
+      "@utils": ["./utils"]
+    }
+  },
+  "exclude": ["node_modules"]
+}
