+++ src/components/Keypad/Keypad/KeypadFC.jsx
@@ -1,21 +1,22 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import BUTTONS from '@constants/buttons';
+import controller from '@utils/controller';
 
 import Button from '../Button/ButtonFC';
 
 import WrapperKeypad from './styled';
 
-const Keypad = ({ handle }) => {
-    const actualDigit = useSelector(({ control: { actualDigit: act } }) => act);
+const Keypad = () => {
+    const dispatch = useDispatch();
     return (
-        <WrapperKeypad area={actualDigit}>
+        <WrapperKeypad>
             {BUTTONS.map(({ name, digit }) => (
                 <Button
                     key={name}
                     name={name}
                     area={name}
-                    onClick={handle(digit, name)}
+                    onClick={controller(digit, name, dispatch)}
                 >
                     {digit}
                 </Button>
